Moved component from services->components
